refactor(NcSettingsSection): migrate to Typescript and script-setup
