+++ src/components/index.ts
@@ -76,7 +76,7 @@ export { default as NcSelect } from './NcSelect/index.js'
 export { default as NcSelectTags } from './NcSelectTags/index.js'
 export { default as NcSelectUsers } from './NcSelectUsers/index.ts'
 export { default as NcSettingsInputText } from './NcSettingsInputText/index.js'
-export { default as NcSettingsSection } from './NcSettingsSection/index.js'
+export { default as NcSettingsSection } from './NcSettingsSection/index.ts'
 export { default as NcSettingsSelectGroup } from './NcSettingsSelectGroup/index.js'
 export { default as NcTextArea } from './NcTextArea/index.ts'
 export { default as NcTextField } from './NcTextField/index.ts'
